+++ source/Examples/Resources/CMCollectionMembershipEvaluationComponent/CMCollectionMembershipEvaluationComponent_Example.ps1
@@ -1,0 +1,17 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to configure the Collection Membership Evaluation Component in Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMCollectionMembershipEvaluationComponent ExampleSettings
+        {
+            SiteCode       = 'LAB'
+            EvaluationMins = 5
+        }
+    }
+}